webscraper/requirements.txt
# Html parsing
selenium==4.40.0
beautifulsoup4==4.14.3
nh3==0.3.3

# Http
aiohttp==3.13.3
fake-useragent==2.2.0

# Data handling
pandas==3.0.1
pydantic==2.12.5
pydantic-settings==2.13.0

# Utils
tqdm==4.67.3
pyyaml==6.0.3
openpyxl==3.1.5
aiofiles==25.1.0
